+++ snowchains_core/CHANGELOG.md
@@ -1,10 +1,10 @@
 # Changelog
 
-## [0.5.1] - 2020-09-24Z
+## [Unreleased]
 
 ### Fixed
 
-- Supports "小数誤差許容問題" in yukicoder.
+- Supports "小数誤差許容問題" in yukicoder. ([#96](https://github.com/qryxip/snowchains/pull/96))
 
 ## [0.5.0] - 2020-08-25Z